refactor: make public layout into the default layout
