+++ src/layouts/default/DefaultFooter.tsx
@@ -1,6 +1,6 @@
 import { Box } from '@mui/material';
 
-export function PublicFooter() {
+export function DefaultFooter() {
   return (
     <Box textAlign="center" marginY={4}>
       Seasoning © {new Date().getFullYear()}
